requirements.txt
# Основные зависимости для анализа графиков и генерации отчётов

# Обработка данных
pandas>=1.5.0
numpy>=1.21.0

# Визуализация
matplotlib>=3.5.0
seaborn>=0.11.0

# Получение данных
yfinance>=0.2.0
ccxt>=3.0.0

# Научные вычисления
scipy>=1.9.0

# Обработка конфигурации
PyYAML>=6.0

# Генерация отчётов
Jinja2>=3.1.0

# Дополнительные утилиты
requests>=2.28.0
python-dateutil>=2.8.0
